requirements.txt
streamlit
requestsc